requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-bootstrap-breadcrumbs==0.9.2
django-dynamic-model-validation==0.1.13
django-extensions==2.2.9
django-grappelli==2.14.2
djangorestframework==3.11.0
future==0.17.1
idna==2.9
isodate==0.6.0
lxml==4.5.1
mysql-connector-python==8.0.20
networkx==2.4
oauthlib==3.1.0
protobuf==3.12.2
prov==1.5.3
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
rdflib==5.0.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
